saving progress on directions
